DCAC-68: Refactor to cut test execution time.
